Implemented service discovery
