+++ app/src/main/java/com/group12/activities/WiFiDirectActivity.java
@@ -1,377 +1,375 @@
-
-package com.group12.activities;
-
-import android.Manifest;
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-import android.content.IntentFilter;
-import android.content.pm.PackageManager;
-import android.net.wifi.p2p.WifiP2pConfig;
-import android.net.wifi.p2p.WifiP2pDevice;
-import android.net.wifi.p2p.WifiP2pManager;
-import android.net.wifi.p2p.WifiP2pManager.ActionListener;
-import android.net.wifi.p2p.WifiP2pManager.Channel;
-import android.net.wifi.p2p.WifiP2pManager.ChannelListener;
-import android.net.wifi.p2p.WifiP2pManager.DnsSdServiceResponseListener;
-import android.net.wifi.p2p.nsd.WifiP2pDnsSdServiceInfo;
-import android.net.wifi.p2p.nsd.WifiP2pDnsSdServiceRequest;
-import android.os.Build;
-import android.os.Bundle;
-import android.provider.Settings;
-import android.util.Log;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.TextView;
-import android.widget.Toast;
-
-
-import androidx.appcompat.app.AppCompatActivity;
-import com.group12.p2p.DeviceDetailFragment;
-import com.group12.p2p.DeviceListFragment;
-import com.group12.p2p.WiFiDirectBroadcastReceiver;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * An activity that uses WiFi Direct APIs to discover and connect with available
- * devices. WiFi Direct APIs are asynchronous and rely on callback mechanism
- * using interfaces to notify the application of operation success or failure.
- * The application should also register a BroadcastReceiver for notification of
- * WiFi state related events.
- */
-public class WiFiDirectActivity extends AppCompatActivity implements ChannelListener, DeviceListFragment.DeviceActionListener {
-
-    public static final String TAG = "WifiServiceDiscovery";
-    public static final String SERVICE_INSTANCE = "wayfinder";
-    public static final String TXTRECORD_PROP_AVAILABLE = "available";
-    public static final String SERVICE_REG_TYPE = "_presence._tcp";
-
-    private static final int PERMISSIONS_REQUEST_CODE_ACCESS_FINE_LOCATION = 1001;
-
-    private WifiP2pManager manager;
-    private boolean isWifiP2pEnabled = false;
-    private boolean retryChannel = false;
-
-
-    private final IntentFilter intentFilter = new IntentFilter();
-    private Channel channel;
-    private BroadcastReceiver receiver = null;
-
-
-    /**
-     * @param isWifiP2pEnabled the isWifiP2pEnabled to set
-     */
-    public void setIsWifiP2pEnabled(boolean isWifiP2pEnabled) {
-        this.isWifiP2pEnabled = isWifiP2pEnabled;
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String[] permissions,
-            int[] grantResults) {
-        switch (requestCode) {
-        case PERMISSIONS_REQUEST_CODE_ACCESS_FINE_LOCATION:
-            if  (grantResults[0] != PackageManager.PERMISSION_GRANTED) {
-                Log.e(TAG, "Fine location permission is not granted!");
-                finish();
-            }
-            break;
-        }
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_p2p);
-
-        // add necessary intent values to be matched.
-
-        intentFilter.addAction(WifiP2pManager.WIFI_P2P_STATE_CHANGED_ACTION);
-        intentFilter.addAction(WifiP2pManager.WIFI_P2P_PEERS_CHANGED_ACTION);
-        intentFilter.addAction(WifiP2pManager.WIFI_P2P_CONNECTION_CHANGED_ACTION);
-        intentFilter.addAction(WifiP2pManager.WIFI_P2P_THIS_DEVICE_CHANGED_ACTION);
-
-        manager = (WifiP2pManager) getSystemService(Context.WIFI_P2P_SERVICE);
-        channel = manager.initialize(this, getMainLooper(), null);
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M
-                    && checkSelfPermission(Manifest.permission.ACCESS_FINE_LOCATION)
-                    != PackageManager.PERMISSION_GRANTED) {
-            requestPermissions(new String[]{Manifest.permission.ACCESS_FINE_LOCATION},
-                    WiFiDirectActivity.PERMISSIONS_REQUEST_CODE_ACCESS_FINE_LOCATION);
-            // After this point you wait for callback in
-            // onRequestPermissionsResult(int, String[], int[]) overridden method
-        }
-
-        receiver = new WiFiDirectBroadcastReceiver(manager, channel, this);
-        registerReceiver(receiver, intentFilter);
-        startRegistrationAndDiscovery();
-    }
-
-
-    private void startRegistrationAndDiscovery() {
-        Map<String, String> record = new HashMap<String, String>();
-        record.put(TXTRECORD_PROP_AVAILABLE, "visible");
-        WifiP2pDnsSdServiceInfo service = WifiP2pDnsSdServiceInfo.newInstance(
-                SERVICE_INSTANCE, SERVICE_REG_TYPE, record);
-        manager.addLocalService(channel, service, new ActionListener() {
-            @Override
-            public void onSuccess() {
-                Toast.makeText(WiFiDirectActivity.this,"Added Local Service",Toast.LENGTH_LONG);
-
-            }
-            @Override
-            public void onFailure(int error) {
-                Toast.makeText(WiFiDirectActivity.this,"Failed to add a service",Toast.LENGTH_LONG);
-            }
-        });
-
-    }
-
-
-
-    /** register the BroadcastReceiver with the intent values to be matched */
-    @Override
-    public void onResume() {
-        super.onResume();
-        receiver = new WiFiDirectBroadcastReceiver(manager, channel, this);
-        registerReceiver(receiver, intentFilter);
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        unregisterReceiver(receiver);
-    }
-
-    /**
-     * Remove all peers and clear all fields. This is called on
-     * BroadcastReceiver receiving a state change event.
-     */
-    public void resetData() {
-        DeviceListFragment fragmentList = (DeviceListFragment) getFragmentManager()
-                .findFragmentById(R.id.frag_list);
-        DeviceDetailFragment fragmentDetails = (DeviceDetailFragment) getFragmentManager()
-                .findFragmentById(R.id.frag_detail);
-        if (fragmentList != null) {
-            fragmentList.clearPeers();
-        }
-        if (fragmentDetails != null) {
-            fragmentDetails.resetViews();
-        }
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.action_items, menu);
-        return true;
-    }
-
-    private void startDiscovery(){
-        // After attaching listeners, create a service request and initiate
-        // discovery.
-        final DeviceListFragment fragment = (DeviceListFragment) getFragmentManager()
-                .findFragmentById(R.id.frag_list);
-        fragment.onInitiateDiscovery();
-        manager.setDnsSdResponseListeners(channel, new DnsSdServiceResponseListener() {
-
-                    @Override
-                    public void onDnsSdServiceAvailable(String instanceName,
-                                                        String registrationType, WifiP2pDevice srcDevice) {
-                        if (instanceName.equalsIgnoreCase(SERVICE_INSTANCE)) {
-                            if (fragment != null) {
-                                fragment.addToList(srcDevice);
-                                Log.d(TAG, "onBonjourServiceAvailable "
-                                        + instanceName);
-                            }
-                        }
-                    }
-                }, new WifiP2pManager.DnsSdTxtRecordListener() {
-                    /**
-                     * A new TXT record is available. Pick up the advertised
-                     * buddy name.
-                     */
-                    @Override
-                    public void onDnsSdTxtRecordAvailable(
-                            String fullDomainName, Map<String, String> record,
-                            WifiP2pDevice device) {
-                        Log.d(TAG,
-                                device.deviceName + " is "
-                                        + record.get(TXTRECORD_PROP_AVAILABLE));
-                    }
-                });
-        WifiP2pDnsSdServiceRequest serviceRequest = WifiP2pDnsSdServiceRequest.newInstance();
-        manager.addServiceRequest(channel, serviceRequest,
-                new ActionListener() {
-                    @Override
-                    public void onSuccess() {
-                        Toast.makeText(WiFiDirectActivity.this,"Added service discovery request",Toast.LENGTH_LONG);
-                    }
-                    @Override
-                    public void onFailure(int arg0) {
-                        Toast.makeText(WiFiDirectActivity.this,"Failed adding service discovery request",Toast.LENGTH_LONG);
-                    }
-                });
-        manager.discoverServices(channel, new ActionListener() {
-            @Override
-            public void onSuccess() {
-                Toast.makeText(WiFiDirectActivity.this,"Added service discovery request",Toast.LENGTH_LONG);
-            }
-
-            @Override
-            public void onFailure(int reason) {
-
-            }
-        });
-    }
-    /*
-     * (non-Javadoc)
-     * @see android.app.Activity#onOptionsItemSelected(android.view.MenuItem)
-     */
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.atn_direct_enable:
-                if (manager != null && channel != null) {
-
-                    // Since this is the system wireless settings activity, it's
-                    // not going to send us a result. We will be notified by
-                    // WiFiDeviceBroadcastReceiver instead.
-
-                    startActivity(new Intent(Settings.ACTION_WIRELESS_SETTINGS));
-                } else {
-                    Log.e(SERVICE_INSTANCE, "channel or manager is null");
-                }
-                return true;
-
-            case R.id.atn_direct_discover:
-                if (!isWifiP2pEnabled) {
-                    Toast.makeText(WiFiDirectActivity.this, R.string.p2p_off_warning,
-                            Toast.LENGTH_SHORT).show();
-                    return true;
-                }
-                startDiscovery();
-
-//                final DeviceListFragment fragment = (DeviceListFragment) getFragmentManager()
-//                        .findFragmentById(R.id.frag_list);
-//                fragment.onInitiateDiscovery();
-//                manager.discoverServices(channel, new WifiP2pManager.ActionListener() {
+//
+//package com.group12.activities;
+//
+//import android.Manifest;
+//import android.content.BroadcastReceiver;
+//import android.content.Context;
+//import android.content.Intent;
+//import android.content.IntentFilter;
+//import android.content.pm.PackageManager;
+//import android.net.wifi.p2p.WifiP2pConfig;
+//import android.net.wifi.p2p.WifiP2pDevice;
+//import android.net.wifi.p2p.WifiP2pManager;
+//import android.net.wifi.p2p.WifiP2pManager.ActionListener;
+//import android.net.wifi.p2p.WifiP2pManager.Channel;
+//import android.net.wifi.p2p.WifiP2pManager.ChannelListener;
+//import android.net.wifi.p2p.WifiP2pManager.DnsSdServiceResponseListener;
+//import android.net.wifi.p2p.nsd.WifiP2pDnsSdServiceInfo;
+//import android.net.wifi.p2p.nsd.WifiP2pDnsSdServiceRequest;
+//import android.os.Build;
+//import android.os.Bundle;
+//import android.provider.Settings;
+//import android.util.Log;
+//import android.view.Menu;
+//import android.view.MenuInflater;
+//import android.view.MenuItem;
+//import android.view.View;
+//import android.widget.Toast;
+//
+//
+//import androidx.appcompat.app.AppCompatActivity;
+//import com.group12.p2p.DeviceListFragment;
+//import com.group12.p2p.WiFiDirectBroadcastReceiver;
+//
+//import java.util.HashMap;
+//import java.util.Map;
+//
+///**
+// * An activity that uses WiFi Direct APIs to discover and connect with available
+// * devices. WiFi Direct APIs are asynchronous and rely on callback mechanism
+// * using interfaces to notify the application of operation success or failure.
+// * The application should also register a BroadcastReceiver for notification of
+// * WiFi state related events.
+// */
+//public class WiFiDirectActivity extends AppCompatActivity implements ChannelListener, DeviceListFragment.DeviceActionListener {
+//
+//    public static final String TAG = "WifiServiceDiscovery";
+//    public static final String SERVICE_INSTANCE = "wayfinder";
+//    public static final String TXTRECORD_PROP_AVAILABLE = "available";
+//    public static final String SERVICE_REG_TYPE = "_presence._tcp";
+//
+//    private static final int PERMISSIONS_REQUEST_CODE_ACCESS_FINE_LOCATION = 1001;
+//
+//    private WifiP2pManager manager;
+//    private boolean isWifiP2pEnabled = false;
+//    private boolean retryChannel = false;
+//
+//
+//    private final IntentFilter intentFilter = new IntentFilter();
+//    private Channel channel;
+//    private BroadcastReceiver receiver = null;
+//
+//
+//    /**
+//     * @param isWifiP2pEnabled the isWifiP2pEnabled to set
+//     */
+//    public void setIsWifiP2pEnabled(boolean isWifiP2pEnabled) {
+//        this.isWifiP2pEnabled = isWifiP2pEnabled;
+//    }
+//
+//    @Override
+//    public void onRequestPermissionsResult(int requestCode, String[] permissions,
+//            int[] grantResults) {
+//        switch (requestCode) {
+//        case PERMISSIONS_REQUEST_CODE_ACCESS_FINE_LOCATION:
+//            if  (grantResults[0] != PackageManager.PERMISSION_GRANTED) {
+//                Log.e(TAG, "Fine location permission is not granted!");
+//                finish();
+//            }
+//            break;
+//        }
+//    }
+//
+//    @Override
+//    public void onCreate(Bundle savedInstanceState) {
+//        super.onCreate(savedInstanceState);
+//        setContentView(R.layout.activity_p2p);
+//
+//        // add necessary intent values to be matched.
+//
+//        intentFilter.addAction(WifiP2pManager.WIFI_P2P_STATE_CHANGED_ACTION);
+//        intentFilter.addAction(WifiP2pManager.WIFI_P2P_PEERS_CHANGED_ACTION);
+//        intentFilter.addAction(WifiP2pManager.WIFI_P2P_CONNECTION_CHANGED_ACTION);
+//        intentFilter.addAction(WifiP2pManager.WIFI_P2P_THIS_DEVICE_CHANGED_ACTION);
+//
+//        manager = (WifiP2pManager) getSystemService(Context.WIFI_P2P_SERVICE);
+//        channel = manager.initialize(this, getMainLooper(), null);
+//
+//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M
+//                    && checkSelfPermission(Manifest.permission.ACCESS_FINE_LOCATION)
+//                    != PackageManager.PERMISSION_GRANTED) {
+//            requestPermissions(new String[]{Manifest.permission.ACCESS_FINE_LOCATION},
+//                    WiFiDirectActivity.PERMISSIONS_REQUEST_CODE_ACCESS_FINE_LOCATION);
+//            // After this point you wait for callback in
+//            // onRequestPermissionsResult(int, String[], int[]) overridden method
+//        }
+//
+//        //receiver = new WiFiDirectBroadcastReceiver(manager, channel, this);
+//        registerReceiver(receiver, intentFilter);
+//        startRegistrationAndDiscovery();
+//    }
+//
+//
+//    private void startRegistrationAndDiscovery() {
+//        Map<String, String> record = new HashMap<String, String>();
+//        record.put(TXTRECORD_PROP_AVAILABLE, "visible");
+//        WifiP2pDnsSdServiceInfo service = WifiP2pDnsSdServiceInfo.newInstance(
+//                SERVICE_INSTANCE, SERVICE_REG_TYPE, record);
+//        manager.addLocalService(channel, service, new ActionListener() {
+//            @Override
+//            public void onSuccess() {
+//                Toast.makeText(WiFiDirectActivity.this,"Added Local Service",Toast.LENGTH_LONG);
+//
+//            }
+//            @Override
+//            public void onFailure(int error) {
+//                Toast.makeText(WiFiDirectActivity.this,"Failed to add a service",Toast.LENGTH_LONG);
+//            }
+//        });
+//
+//    }
+//
+//
+//
+//    /** register the BroadcastReceiver with the intent values to be matched */
+//    @Override
+//    public void onResume() {
+//        super.onResume();
+//        //receiver = new WiFiDirectBroadcastReceiver(manager, channel, this);
+//        registerReceiver(receiver, intentFilter);
+//    }
+//
+//    @Override
+//    public void onPause() {
+//        super.onPause();
+//        unregisterReceiver(receiver);
+//    }
+//
+//    /**
+//     * Remove all peers and clear all fields. This is called on
+//     * BroadcastReceiver receiving a state change event.
+//     */
+////    public void resetData() {
+////        DeviceListFragment fragmentList = (DeviceListFragment) getFragmentManager()
+////                .findFragmentById(R.id.frag_list);
+////        DeviceDetail fragmentDetails = (DeviceDetail) getFragmentManager()
+////                .findFragmentById(R.id.frag_detail);
+////        if (fragmentList != null) {
+////            fragmentList.clearPeers();
+////        }
+////        if (fragmentDetails != null) {
+////            fragmentDetails.resetViews();
+////        }
+////    }
+//
+//    @Override
+//    public boolean onCreateOptionsMenu(Menu menu) {
+//        MenuInflater inflater = getMenuInflater();
+//        inflater.inflate(R.menu.action_items, menu);
+//        return true;
+//    }
+//
+//    private void startDiscovery(){
+//        // After attaching listeners, create a service request and initiate
+//        // discovery.
+//        final DeviceListFragment fragment = (DeviceListFragment) getFragmentManager()
+//                .findFragmentById(R.id.frag_list);
+//        fragment.onInitiateDiscovery();
+//        manager.setDnsSdResponseListeners(channel, new DnsSdServiceResponseListener() {
+//
+//                    @Override
+//                    public void onDnsSdServiceAvailable(String instanceName,
+//                                                        String registrationType, WifiP2pDevice srcDevice) {
+//                        if (instanceName.equalsIgnoreCase(SERVICE_INSTANCE)) {
+//                            if (fragment != null) {
+//                                fragment.addToList(srcDevice);
+//                                Log.d(TAG, "onBonjourServiceAvailable "
+//                                        + instanceName);
+//                            }
+//                        }
+//                    }
+//                }, new WifiP2pManager.DnsSdTxtRecordListener() {
+//                    /**
+//                     * A new TXT record is available. Pick up the advertised
+//                     * buddy name.
+//                     */
+//                    @Override
+//                    public void onDnsSdTxtRecordAvailable(
+//                            String fullDomainName, Map<String, String> record,
+//                            WifiP2pDevice device) {
+//                        Log.d(TAG,
+//                                device.deviceName + " is "
+//                                        + record.get(TXTRECORD_PROP_AVAILABLE));
+//                    }
+//                });
+//        WifiP2pDnsSdServiceRequest serviceRequest = WifiP2pDnsSdServiceRequest.newInstance();
+//        manager.addServiceRequest(channel, serviceRequest,
+//                new ActionListener() {
+//                    @Override
+//                    public void onSuccess() {
+//                        Toast.makeText(WiFiDirectActivity.this,"Added service discovery request",Toast.LENGTH_LONG);
+//                    }
+//                    @Override
+//                    public void onFailure(int arg0) {
+//                        Toast.makeText(WiFiDirectActivity.this,"Failed adding service discovery request",Toast.LENGTH_LONG);
+//                    }
+//                });
+//        manager.discoverServices(channel, new ActionListener() {
+//            @Override
+//            public void onSuccess() {
+//                Toast.makeText(WiFiDirectActivity.this,"Added service discovery request",Toast.LENGTH_LONG);
+//            }
+//
+//            @Override
+//            public void onFailure(int reason) {
+//
+//            }
+//        });
+//    }
+//    /*
+//     * (non-Javadoc)
+//     * @see android.app.Activity#onOptionsItemSelected(android.view.MenuItem)
+//     */
+//    @Override
+//    public boolean onOptionsItemSelected(MenuItem item) {
+//        switch (item.getItemId()) {
+//            case R.id.atn_direct_enable:
+//                if (manager != null && channel != null) {
+//
+//                    // Since this is the system wireless settings activity, it's
+//                    // not going to send us a result. We will be notified by
+//                    // WiFiDeviceBroadcastReceiver instead.
+//
+//                    startActivity(new Intent(Settings.ACTION_WIRELESS_SETTINGS));
+//                } else {
+//                    Log.e(SERVICE_INSTANCE, "channel or manager is null");
+//                }
+//                return true;
+//
+//            case R.id.atn_direct_discover:
+//                if (!isWifiP2pEnabled) {
+//                    Toast.makeText(WiFiDirectActivity.this, R.string.p2p_off_warning,
+//                            Toast.LENGTH_SHORT).show();
+//                    return true;
+//                }
+//                startDiscovery();
+//
+////                final DeviceListFragment fragment = (DeviceListFragment) getFragmentManager()
+////                        .findFragmentById(R.id.frag_list);
+////                fragment.onInitiateDiscovery();
+////                manager.discoverServices(channel, new WifiP2pManager.ActionListener() {
+////
+////                    @Override
+////                    public void onSuccess() {
+////                        Toast.makeText(WiFiDirectActivity.this, "Discovery Initiated",
+////                                Toast.LENGTH_SHORT).show();
+////                    }
+////
+////                    @Override
+////                    public void onFailure(int reasonCode) {
+////                        Toast.makeText(WiFiDirectActivity.this, "Discovery Failed : " + reasonCode,
+////                                Toast.LENGTH_SHORT).show();
+////                    }
+////                });
+//                return true;
+//            default:
+//                return super.onOptionsItemSelected(item);
+//        }
+//    }
+//
+////    @Override
+////    public void showDetails(WifiP2pDevice device) {
+////        DeviceDetail fragment = (DeviceDetail) getFragmentManager()
+////                .findFragmentById(R.id.frag_detail);
+////        fragment.showDetails(device);
+////    }
+//
+//    @Override
+//    public void connect(WifiP2pConfig config) {
+//        manager.connect(channel, config, new ActionListener() {
+//
+//            @Override
+//            public void onSuccess() {
+//                // WiFiDirectBroadcastReceiver will notify us. Ignore for now.
+//            }
+//
+//            @Override
+//            public void onFailure(int reason) {
+//                Toast.makeText(WiFiDirectActivity.this, "Connect failed. Retry.",
+//                        Toast.LENGTH_SHORT).show();
+//            }
+//        });
+//    }
+//
+////    @Override
+////    public void disconnect() {
+////        final DeviceDetail fragment = (DeviceDetail) getFragmentManager()
+////                .findFragmentById(R.id.frag_detail);
+////        fragment.resetViews();
+////        manager.removeGroup(channel, new ActionListener() {
+////
+////            @Override
+////            public void onFailure(int reasonCode) {
+////                Log.d(TAG, "Disconnect failed. Reason :" + reasonCode);
+////
+////            }
+////            @Override
+////            public void onSuccess() {
+////                fragment.getView().setVisibility(View.GONE);
+////            }
+////        });
+////    }
+//
+////    @Override
+////    public void onChannelDisconnected() {
+////        // we will try once more
+////        if (manager != null && !retryChannel) {
+////            Toast.makeText(this, "Channel lost. Trying again", Toast.LENGTH_LONG).show();
+////            resetData();
+////            retryChannel = true;
+////            manager.initialize(this, getMainLooper(), this);
+////        } else {
+////            Toast.makeText(this,
+////                    "Severe! Channel is probably lost premanently. Try Disable/Re-Enable P2P.",
+////                    Toast.LENGTH_LONG).show();
+////        }
+////    }
+//
+//    @Override
+//    public void cancelDisconnect() {
+//
+//        /*
+//         * A cancel abort request by user. Disconnect i.e. removeGroup if
+//         * already connected. Else, request WifiP2pManager to abort the ongoing
+//         * request
+//         */
+//        if (manager != null) {
+//            final DeviceListFragment fragment = (DeviceListFragment) getFragmentManager()
+//                    .findFragmentById(R.id.frag_list);
+//            if (fragment.getDevice() == null
+//                    || fragment.getDevice().status == WifiP2pDevice.CONNECTED) {
+//                disconnect();
+//            } else if (fragment.getDevice().status == WifiP2pDevice.AVAILABLE
+//                    || fragment.getDevice().status == WifiP2pDevice.INVITED) {
+//
+//                manager.cancelConnect(channel, new ActionListener() {
 //
 //                    @Override
 //                    public void onSuccess() {
-//                        Toast.makeText(WiFiDirectActivity.this, "Discovery Initiated",
+//                        Toast.makeText(WiFiDirectActivity.this, "Aborting connection",
 //                                Toast.LENGTH_SHORT).show();
 //                    }
 //
 //                    @Override
 //                    public void onFailure(int reasonCode) {
-//                        Toast.makeText(WiFiDirectActivity.this, "Discovery Failed : " + reasonCode,
+//                        Toast.makeText(WiFiDirectActivity.this,
+//                                "Connect abort request failed. Reason Code: " + reasonCode,
 //                                Toast.LENGTH_SHORT).show();
 //                    }
 //                });
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-    }
-
-    @Override
-    public void showDetails(WifiP2pDevice device) {
-        DeviceDetailFragment fragment = (DeviceDetailFragment) getFragmentManager()
-                .findFragmentById(R.id.frag_detail);
-        fragment.showDetails(device);
-    }
-
-    @Override
-    public void connect(WifiP2pConfig config) {
-        manager.connect(channel, config, new ActionListener() {
-
-            @Override
-            public void onSuccess() {
-                // WiFiDirectBroadcastReceiver will notify us. Ignore for now.
-            }
-
-            @Override
-            public void onFailure(int reason) {
-                Toast.makeText(WiFiDirectActivity.this, "Connect failed. Retry.",
-                        Toast.LENGTH_SHORT).show();
-            }
-        });
-    }
-
-    @Override
-    public void disconnect() {
-        final DeviceDetailFragment fragment = (DeviceDetailFragment) getFragmentManager()
-                .findFragmentById(R.id.frag_detail);
-        fragment.resetViews();
-        manager.removeGroup(channel, new ActionListener() {
-
-            @Override
-            public void onFailure(int reasonCode) {
-                Log.d(TAG, "Disconnect failed. Reason :" + reasonCode);
-
-            }
-            @Override
-            public void onSuccess() {
-                fragment.getView().setVisibility(View.GONE);
-            }
-        });
-    }
-
-    @Override
-    public void onChannelDisconnected() {
-        // we will try once more
-        if (manager != null && !retryChannel) {
-            Toast.makeText(this, "Channel lost. Trying again", Toast.LENGTH_LONG).show();
-            resetData();
-            retryChannel = true;
-            manager.initialize(this, getMainLooper(), this);
-        } else {
-            Toast.makeText(this,
-                    "Severe! Channel is probably lost premanently. Try Disable/Re-Enable P2P.",
-                    Toast.LENGTH_LONG).show();
-        }
-    }
-
-    @Override
-    public void cancelDisconnect() {
-
-        /*
-         * A cancel abort request by user. Disconnect i.e. removeGroup if
-         * already connected. Else, request WifiP2pManager to abort the ongoing
-         * request
-         */
-        if (manager != null) {
-            final DeviceListFragment fragment = (DeviceListFragment) getFragmentManager()
-                    .findFragmentById(R.id.frag_list);
-            if (fragment.getDevice() == null
-                    || fragment.getDevice().status == WifiP2pDevice.CONNECTED) {
-                disconnect();
-            } else if (fragment.getDevice().status == WifiP2pDevice.AVAILABLE
-                    || fragment.getDevice().status == WifiP2pDevice.INVITED) {
-
-                manager.cancelConnect(channel, new ActionListener() {
-
-                    @Override
-                    public void onSuccess() {
-                        Toast.makeText(WiFiDirectActivity.this, "Aborting connection",
-                                Toast.LENGTH_SHORT).show();
-                    }
-
-                    @Override
-                    public void onFailure(int reasonCode) {
-                        Toast.makeText(WiFiDirectActivity.this,
-                                "Connect abort request failed. Reason Code: " + reasonCode,
-                                Toast.LENGTH_SHORT).show();
-                    }
-                });
-            }
-        }
-
-    }
-
-}
+//            }
+//        }
+//
+//    }
+//
+//}
