+++ app/src/main/java/com/group12/activities/WifiDirectService.java
@@ -1,40 +1,41 @@
 package com.group12.activities;
 
 import android.app.IntentService;
-import android.Manifest;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.content.pm.PackageManager;
+import android.net.Uri;
+import android.net.wifi.WpsInfo;
 import android.net.wifi.p2p.WifiP2pConfig;
 import android.net.wifi.p2p.WifiP2pDevice;
+import android.net.wifi.p2p.WifiP2pInfo;
 import android.net.wifi.p2p.WifiP2pManager;
 import android.net.wifi.p2p.WifiP2pManager.ActionListener;
 import android.net.wifi.p2p.WifiP2pManager.Channel;
-import android.net.wifi.p2p.WifiP2pManager.ChannelListener;
 import android.net.wifi.p2p.WifiP2pManager.DnsSdServiceResponseListener;
 import android.net.wifi.p2p.nsd.WifiP2pDnsSdServiceInfo;
 import android.net.wifi.p2p.nsd.WifiP2pDnsSdServiceRequest;
+import android.os.AsyncTask;
 import android.os.Binder;
-import android.os.Build;
-import android.os.Bundle;
 import android.os.IBinder;
-import android.provider.Settings;
 import android.util.Log;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.TextView;
 import android.widget.Toast;
 
 
-import androidx.appcompat.app.AppCompatActivity;
-import com.group12.p2p.DeviceDetailFragment;
-import com.group12.p2p.DeviceListFragment;
+import androidx.core.content.FileProvider;
+
+
 import com.group12.p2p.WiFiDirectBroadcastReceiver;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.net.ServerSocket;
+import java.net.Socket;
+import java.sql.Connection;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
@@ -50,7 +51,7 @@ import java.util.Map;
  * TODO: Customize class - update intent actions, extra parameters and static
  * helper methods.
  */
-public class WifiDirectService extends IntentService {
+public class WifiDirectService extends IntentService implements WifiP2pManager.ConnectionInfoListener {
     public static final String TAG = "WifiServiceDiscovery";
     public static final String SERVICE_INSTANCE = "wayfinder";
     public static final String TXTRECORD_PROP_AVAILABLE = "available";
@@ -62,11 +63,10 @@ public class WifiDirectService extends IntentService {
     private boolean isWifiP2pEnabled = false;
     private boolean retryChannel = false;
 
-
+    private Context mapContext;
     private final IntentFilter intentFilter = new IntentFilter();
     private Channel channel;
     private BroadcastReceiver receiver = null;
-
     private final IBinder myBinder = new MyLocalBinder();
 
 
@@ -78,9 +78,7 @@ public class WifiDirectService extends IntentService {
     }
 
 
-    public WifiDirectService() {
-        super("WifiDirectService");
-    }
+    public WifiDirectService() { super("WifiDirectService"); }
 
 
 
@@ -93,9 +91,8 @@ public class WifiDirectService extends IntentService {
      * Handle action Foo in the provided background thread with the provided
      * parameters.
      */
-    private void handleActionFoo(String param1, String param2) {
-        // TODO: Handle action Foo
-        throw new UnsupportedOperationException("Not yet implemented");
+    public void resetData() {
+        //To do
     }
 
     /**
@@ -118,6 +115,31 @@ public class WifiDirectService extends IntentService {
         return (dateFormat.format(new Date()));
     }
 
+    public void initialiseWifiService(Context context){
+        // add necessary intent values to be matched.
+        mapContext = context.getApplicationContext();
+        intentFilter.addAction(WifiP2pManager.WIFI_P2P_STATE_CHANGED_ACTION);
+        intentFilter.addAction(WifiP2pManager.WIFI_P2P_PEERS_CHANGED_ACTION);
+        intentFilter.addAction(WifiP2pManager.WIFI_P2P_CONNECTION_CHANGED_ACTION);
+        intentFilter.addAction(WifiP2pManager.WIFI_P2P_THIS_DEVICE_CHANGED_ACTION);
+
+        manager = (WifiP2pManager) getSystemService(Context.WIFI_P2P_SERVICE);
+        channel = manager.initialize(this, getMainLooper(), null);
+
+//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M
+//                && checkSelfPermission(Manifest.permission.ACCESS_FINE_LOCATION)
+//                != PackageManager.PERMISSION_GRANTED) {
+//            requestPermissions(new String[]{Manifest.permission.ACCESS_FINE_LOCATION},
+//                    WiFiDirectActivity.PERMISSIONS_REQUEST_CODE_ACCESS_FINE_LOCATION);
+//            // After this point you wait for callback in
+//            // onRequestPermissionsResult(int, String[], int[]) overridden method
+//        }
+
+        receiver = new WiFiDirectBroadcastReceiver(manager, channel, this);
+        registerReceiver(receiver, intentFilter);
+        startRegistration();
+    }
+
     public class MyLocalBinder extends Binder {
         WifiDirectService getService() {
             return WifiDirectService.this;
@@ -138,12 +160,17 @@ public class WifiDirectService extends IntentService {
             @Override
             public void onFailure(int error) {
                 Toast.makeText(WifiDirectService.this,"Failed to add a service",Toast.LENGTH_LONG);
+
             }
         });
-
     }
 
-    private void startDiscovery(){
+    public WifiP2pManager getManager(){
+        return manager;
+    }
+
+
+    public void startDiscovery(){
         // After attaching listeners, create a service request and initiate
         // discovery.
 
@@ -153,7 +180,8 @@ public class WifiDirectService extends IntentService {
             public void onDnsSdServiceAvailable(String instanceName,
                                                 String registrationType, WifiP2pDevice srcDevice) {
                 if (instanceName.equalsIgnoreCase(SERVICE_INSTANCE)) {
-                    peers.add(srcDevice);
+                    //peers.add(srcDevice);
+                    connectAndTransfer(srcDevice);
                     Log.d(TAG, "wayfinderServiceAvailable " + instanceName);
                 }
             }
@@ -191,9 +219,152 @@ public class WifiDirectService extends IntentService {
 
             @Override
             public void onFailure(int reason) {
+                Toast.makeText(WifiDirectService.this,"Added service discovery request failed "+reason,Toast.LENGTH_LONG);
 
             }
         });
+    }
+
+    public void connectAndTransfer(WifiP2pDevice device){
+
+        WifiP2pConfig config = new WifiP2pConfig();
+        config.deviceAddress = device.deviceAddress;
+        config.wps.setup = WpsInfo.PBC;
+
+        manager.connect(channel, config, new ActionListener() {
+
+            @Override
+            public void onSuccess() {
+                Toast.makeText(WifiDirectService.this, "Connect successful.",
+                        Toast.LENGTH_SHORT).show();
+                // WiFiDirectBroadcastReceiver will notify us. Ignore for now.
+            }
+
+            @Override
+            public void onFailure(int reason) {
+                Toast.makeText(WifiDirectService.this, "Connect failed. Retry.",
+                        Toast.LENGTH_SHORT).show();
+            }
+        });
+
+    }
+
+    public void onConnectionInfoAvailable(final WifiP2pInfo info) {
+
+
+        if (info.groupFormed) {
+            new FileServerAsyncTask(mapContext)
+                    .execute();
+        }
+        // hide the connect button
+    }
+
+
+    /**
+     * Updates the UI with device data
+     *
+     * @param device the device to be displayed
+     */
+//    public void showDetails(WifiP2pDevice device) {
+//        this.device = device;
+//        Objects.requireNonNull(this.getView()).setVisibility(View.VISIBLE);
+//        TextView view = (TextView) mContentView.findViewById(R.id.device_address);
+//        view.setText(device.deviceAddress);
+//        view = (TextView) mContentView.findViewById(R.id.device_info);
+//        view.setText(device.toString());
+//
+//    }
+
+    /**
+     * A simple server socket that accepts connection and writes some data on
+     * the stream.
+     */
+    public static class FileServerAsyncTask extends AsyncTask<Void, Void, String> {
+
+        private Context context;
+
+        /**
+         * @param context
+         */
+        public FileServerAsyncTask(Context context) {
+            this.context = context;
+        }
+
+        @Override
+        protected String doInBackground(Void... params) {
+            try {
+                ServerSocket serverSocket = new ServerSocket(8988);
+                Log.d(WifiDirectService.TAG, "Server: Socket opened");
+                Socket client = serverSocket.accept();
+                Log.d(WifiDirectService.TAG, "Server: connection done");
+                final File f = new File(context.getExternalFilesDir("received"),
+                        "wifip2pshared-" + System.currentTimeMillis()
+                                + ".jpg");
+
+                File dirs = new File(f.getParent());
+                if (!dirs.exists())
+                    dirs.mkdirs();
+                f.createNewFile();
+
+                Log.d(WifiDirectService.TAG, "server: copying files " + f.toString());
+                InputStream inputstream = client.getInputStream();
+                copyFile(inputstream, new FileOutputStream(f));
+                serverSocket.close();
+                return f.getAbsolutePath();
+            } catch (IOException e) {
+                Log.e(WifiDirectService.TAG, e.getMessage());
+                return null;
+            }
+        }
+
+        /*
+         * (non-Javadoc)
+         * @see android.os.AsyncTask#onPostExecute(java.lang.Object)
+         */
+        @Override
+        protected void onPostExecute(String result) {
+            if (result != null) {
+
+                File recvFile = new File(result);
+                Uri fileUri = FileProvider.getUriForFile(
+                        context,
+                        "com.group12.p2p.fileprovider",
+                        recvFile);
+                Intent intent = new Intent();
+                intent.setAction(android.content.Intent.ACTION_VIEW);
+                intent.setDataAndType(fileUri, "image/*");
+                intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+                context.startActivity(intent);
+            }
+
+        }
+
+        /*
+         * (non-Javadoc)
+         * @see android.os.AsyncTask#onPreExecute()
+         */
+        @Override
+        protected void onPreExecute() {
+            //
+        }
+
+    }
+
+    public static boolean copyFile(InputStream inputStream, OutputStream out) {
+        byte buf[] = new byte[1024];
+        int len;
+        try {
+            while ((len = inputStream.read(buf)) != -1) {
+                out.write(buf, 0, len);
+
+            }
+            out.close();
+            inputStream.close();
+        } catch (IOException e) {
+            Log.d(WifiDirectService.TAG, e.toString());
+            return false;
+        }
+        return true;
     }
