--- app/src/main/java/com/group12/p2p/DeviceManagementService.java
@@ -1,4 +1,0 @@
-package com.group12.p2p;
-
-public class DeviceManagementService {
-}
