--- app/src/main/res/layout/activity_p2p.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:orientation="vertical"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent">
-    <LinearLayout
-            android:layout_width="match_parent"
-            android:id="@+id/linearLayout1"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-        <fragment
-                class="com.group12.p2p.DeviceListFragment"
-                android:id="@+id/frag_list"
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/phone_list_height">
-            <!-- Preview: layout=@layout/row_devices -->
-        </fragment>
-        <fragment
-                class="com.group12.p2p.DeviceDetailFragment"
-                android:id="@+id/frag_detail"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
-            <!-- Preview: layout=@layout/device_detail -->
-        </fragment>
-    </LinearLayout>
-</LinearLayout>
